pyproject.toml
[project]
name = "chunking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.3",
    "hydra-core>=1.3.2",
    "pandas>=2.2.3",
    "sentence-transformers>=4.0.1",
    "tiktoken>=0.9.0",
]
